jspresso-ce-332 Decimal value with maxFractionDigit is not displaying decimals fixes #332

+++ qooxdoo/server/src/main/java/org/jspresso/framework/qooxdoo/rpc/RemoteCallUtils.java
@@ -74,7 +74,7 @@ public class RemoteCallUtils extends net.sf.qooxdoo.rpc.RemoteCallUtils {
         if (((Double) obj).isNaN() || ((Double) obj).isInfinite()) {
           returnValue = null;
         } else {
-          returnValue = super.fromJava(((Double) obj).longValue());
+          returnValue = super.fromJava(((Double) obj).doubleValue());
         }
       } else {
         returnValue = super.fromJava(obj);
